+++ CHANGELOG.md
@@ -5,8 +5,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Breaking change: Update button `min-width` class on `button-tiny`.
+- Feature: Add `margin` prop to `Title` component.
 
-## [9.4.0]
+## [9.5.0] - 2017-03-30
+
+Features:
+- Add `Marger` component.
+- Add `ButtonImage` component.
+- Add `margin` prop to `Paragraph` component.
+
+Fixes:
+- Display correctly disabled `TextInputWithLimit` component.
+
+## [9.4.0] - 2017-03-28
 
 Feature:
 - Add color utilities.